Redefined retry_interval as long to fix no-automatic-reconnect behaviour Preparations for release 2.0.4

+++ php_memcache.h
@@ -82,7 +82,7 @@ typedef struct mmc {
 	unsigned short			port;
 	long					timeout;
 	long					failed;
-	unsigned int			retry_interval;
+	long					retry_interval;
 	int						persistent;
 	int						status;
 } mmc_t;
